# api/openapi.yaml
openapi: 3.0.3
info:
  title: Vesta Kubernetes API
  description: |
    REST API for the Vesta Kubernetes PaaS platform.
    Manages pipelines, apps, deployments, secrets, and templates.
  version: 0.1.0
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.html
servers:
  - url: /api/v1
    description: Vesta API v1
security:
  - BearerAuth: []
...
paths:
...
  /auth/oauth/{provider}:
    parameters:
      - name: provider
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [Auth]
      summary: Initiate OAuth2 login
      operationId: oauthRedirect
      security: []
      responses:
        "302":
          description: Redirect to OAuth provider
  /auth/tokens:
    post:
      tags: [Auth]
      summary: Create an API token
      operationId: createApiToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
                  description: "Allowed scopes (e.g. deploy, read, admin)"
                expiresIn:
                  type: string
                  description: "Duration (e.g. 90d, 1y)"
      responses:
        "201":
          description: Token created (value shown only once)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTokenCreated"
  /auth/tokens/{tokenId}:
    parameters:
      - name: tokenId
        in: path
        required: true
        schema:
          type: string
    delete:
      tags: [Auth]
      summary: Revoke an API token
      operationId: revokeApiToken
      responses:
        "204":
          description: Token revoked

# ════════════════════════════════════════════════════════════════════════════
# Components
# ════════════════════════════════════════════════════════════════════════════
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    ApiTokenCreated:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        token:
          type: string
          description: The token value. Shown only once at creation time.
        scopes:
          type: array
          items:
            type: string
        expiresAt:
          type: string
          format: date-time

    # ── Metrics schemas ──────────────────────────────────────────────────
...